Fixed some confusing bugs in the plan chest and plan library classes.
- wrong plan chest recipe output/name
- more accurate class name for the planlibrary output slot.
